requirements.txt
django==4.2.3
drf_yasg==1.21.7
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
django-user-agents==0.4.0
djangorestframework-simplejwt==5.2.2
psycopg2
pillow==10.0.0
django-modeltranslation
django-ipware